+++ lib/legion/extensions/privatecore.rb
@@ -3,7 +3,9 @@
 require 'legion/extensions/privatecore/version'
 require 'legion/extensions/privatecore/helpers/boundary'
 require 'legion/extensions/privatecore/helpers/erasure'
+require 'legion/extensions/privatecore/helpers/similarity'
 require 'legion/extensions/privatecore/runners/privatecore'
+require 'legion/extensions/privatecore/runners/embedding_guard'
 
 module Legion
   module Extensions